+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## main
 
-- Drop the inexistent `domains.resetDomainToken` method (dnsimple-node#89)
+- NEW: Added `registrar.getDomainPrices` to retrieve whether a domain is premium and the prices to register, transfer, and renew. (dnsimple/dnsimple-node#88)
+- REMOVED: Drop the inexistent `domains.resetDomainToken` method (dnsimple-node#89)
 
 ## Release 4.3.0